+++ src/DAN/server/DanClass.java
@@ -58,7 +58,7 @@ public class DANClass {
 		this.DE_TIME = date;
 	}
 
-	public double getNEUTRON_COUNT() {
+	public int getNEUTRON_COUNT() {
 		return NEUTRON_COUNT;
 	}
